[mtl] Redesign the way descriptor sets are bound.
We used to iterate the individual descriptor layout and checking
for their contained data, multiple times per stage. Instead, we
are now iterating the descriptor pool sub-ranges directly.
Data-wise, this required adding the shader stage mask to the pool data,
and filling it out in allocate_set for immutable samplers.